requirements.txt
geopandas==1.1.0
google-genai==1.19.0
gymnasium==1.1.1
matplotlib==3.10.3
numpy==2.2.6
openai==1.84.0
pandas==2.3.0
pydantic==2.11.5
Requests==2.32.3
Shapely==2.1.1
tiktoken==0.9.0
tqdm==4.67.1
wandb==0.20.1
